+++ Tutorials/01-BasicConcepts/01E-ActionSubscriber/ActionSubscriber/ApiObjects/CustomerEdit.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace BasicConcepts.ActionSubscriber.ApiObjects
+{
+	public class CustomerEdit
+	{
+		public int Id { get; set; } = 42;
+		public byte[] RowVersion { get; set; } = Array.Empty<byte>();
+		public string Name { get; set; }
+
+		public CustomerEdit() { }
+		public CustomerEdit(int id, byte[] rowVersion, string name)
+		{
+			Id = id;
+			RowVersion = rowVersion;
+			Name = name;
+		}
+	}
+}