+++ Tutorials/01-BasicConcepts/01E-ActionSubscriber/ActionSubscriber/Store/EditCustomerUseCase/Feature.cs
@@ -1,0 +1,12 @@
+﻿using Fluxor;
+
+namespace BasicConcepts.ActionSubscriber.Store.EditCustomerUseCase
+{
+	public class Feature : Feature<EditCustomerState>
+	{
+		public override string GetName() => "EditCustomer";
+
+		protected override EditCustomerState GetInitialState() =>
+			new EditCustomerState(isLoading: false);
+	}
+}